% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrdens.R
\name{fit_density}
\alias{fit_density}
\title{Fit Density}
\usage{
fit_density(d)
}
\arguments{
\item{d}{Density object (returned from running \code{density()})}
}
\value{
List of functions \code{ddens()}, \code{pdens()}, \code{qdens()}, \code{rdens()}
}
\description{
Smoothes output of \code{density()} and returns functions needed to run other
functions on the distribution.
}
\examples{
set.seed(123)
data <- rnorm(10000)
dens <- density(data)

fdens <- fit_density(dens)
}
