% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqrdens.R
\name{dens}
\alias{dens}
\alias{ddens}
\alias{pdens}
\alias{qdens}
\alias{rdens}
\title{Arbitrary Distribution}
\usage{
ddens(sdens, x)

pdens(sdens, q)

qdens(sdens, p)

rdens(sdens, n)
}
\arguments{
\item{sdens}{List outputted by \code{smooth_density()}.}

\item{x}{vector of quantiles.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities (should be between 0 and 1).}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function, and random generation for
the \code{density()} object. Each function calls the appropriate function from the
output of \code{smooth_density()}.
}
\examples{
set.seed(123)
data <- rnorm(10000)
dens <- density(data)

sdens <- smooth_density(dens)

ddens(sdens, 2)
pdens(sdens, 2)
qdens(sdens, 0.977)
rdens(sdens, 10L)

}
\seealso{
\verb{stats::\{dpqr\}*()} distribution functions and logspline package.
}
